Review fixes for S2/S6/S3 + E1/scorer VMEM flags + spec 5.2 tier-1 map (V34a/d/e)
Independent review of 5c44acd found five defects; all verified and fixed:
1. -kv_bw_pct capped each UNIT against the DEVICE plate, so two MXUs
   streamed 2x the intended KV rate (P=50 cost only +7%). The budget is
   now one chip-wide per-cycle token bucket (kv_issue_rate / kv_budget_acc,
   refilled in Arch.cc, drawn in State::enqueue_reads, one cycle of carry).
   V34a: P=50 now costs +74% on the KV-dominated shape, CMDs invariant.
2. P=99 vs P=100: the plate cap is not a no-op -- -dram_enq 32 lets a KV
   stream burst 2.2x the plate into the FIFO ahead of weight fetches, and
   pacing at the plate measured 5.6% FASTER. Default kept bit-identical;
   V34a asserts the step < 10%; spec 6 flags -dram_enq = plate for the fit.
3. The reviewer's 'kv jobs still act-prefetched' finding was applied and
   then reverted with a corrected comment: only the KV (weight-side)
   gather is un-prefetchable; the Q panel is an ordinary activation.
4. VecUnitState::init computed byte amounts in int; with S3's batch*seq
   rows an unfused elementwise op over 2^29 elements wrapped its read to 1
   beat. state_transfer and the VPU sites now carry int64 bytes. V34e
   pins the unfused-minus-fused traffic delta on that shape beat-exact
   (51249152; pre-fix 17694720).
5. MEM demand-idle now counts the -data_overhead cycles (V34d +777 exact).

E1 probe keeps VMEM-resident rows flagged instead of dropping them;
score_matrix.py scores rows moving < 150 MB as E1v (VMEM carry, no HBM
verdict). Spec 5.2 records the first tier-1 fidelity map and its five
mechanisms. Suites 67/67 + 5/5.

+++ include/State.h
@@ -133,7 +133,9 @@ struct State {
   void enqueue_reads();
   void check_idle_from_memory();
   bool process_stage();
-  void state_transfer(int st, int read_amt, int write_amt, int min_cycles);
+  // Byte amounts are 64-bit: a batched-prefill elementwise job can exceed
+  // 2^31 bytes (review finding V34e); beats stay int (< 2^31 for < 128 GiB).
+  void state_transfer(int st, int64_t read_amt, int64_t write_amt, int min_cycles);
   virtual void init() = 0;
   virtual bool increment(const enqueue_job_f_t &, int &total_idle, int *n_idle_units) = 0;
   virtual void set_state(int st) = 0;
